requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
djangorestframework-simplejwt==5.2.2